requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
django-environ==0.11.2
selenium==4.12.0
webdriver-manager==4.0.1
django
djangorestframework
gunicorn
dj-database-url
psycopg2-binary
whitenoise[brotli]
